backend/pyproject.toml
[tool.poetry]
name = "football"
version = "0.1.0"
description = ""
authors = ["benjamin oertel <benjamin.oertel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sentry-sdk = "^1.1.0"
dj-database-url = "^0.5.0"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.4"
python-decouple = "^3.4"
django-rq = "^2.4.1"
django-anymail = {extras = ["postmark"], version = "^8.3"}
django-storages = "^1.11.1"
boto3 = "^1.17.84"
django-extensions = "^3.1.3"
ipython = "^7.24.1"

[tool.poetry.dev-dependencies]
black = "^21.5b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
